requirements.txt
django
requests
pandas
pyarrow
pyspark
psycopg2-binary
python-dotenv